+++ api/server/controllers/AuthController.spec.js
@@ -45,6 +45,7 @@ jest.mock('~/models', () => ({
   findSession: jest.fn(),
   updateUser: jest.fn(),
   findUser: jest.fn(),
+  deleteTokens: jest.fn(),
 }));
 jest.mock('~/server/services/RefreshTokenBridge', () => ({
   OPENID_REFRESH_BRIDGE_GRACE_MS: 60 * 1000,
@@ -119,7 +120,11 @@ const openIdClient = require('openid-client');
 const jwt = require('jsonwebtoken');
 const { logger } = require('@librechat/data-schemas');
 const { isEnabled, findOpenIDUser, buildOpenIDRefreshParams } = require('@librechat/api');
-const { graphTokenController, refreshController } = require('./AuthController');
+const {
+  graphTokenController,
+  refreshController,
+  registrationController,
+} = require('./AuthController');
 const { getGraphApiToken } = require('~/server/services/GraphTokenService');
 const {
   clearOpenIDAuthTokens,
@@ -128,9 +133,10 @@ const {
   storeOpenIDSession,
   setCloudFrontAuthCookies,
   setAuthTokens,
+  registerUser,
 } = require('~/server/services/AuthService');
 const { getOpenIdConfig, getOpenIdEmail } = require('~/strategies');
-const { deleteSession, getUserById, findSession, updateUser } = require('~/models');
+const { deleteSession, getUserById, findSession, updateUser, deleteTokens } = require('~/models');
 const {
   createRefreshTokenBridgeFlightKey,
   deleteRefreshTokenBridges,
@@ -2359,5 +2365,79 @@ describe('refreshController – LibreChat path', () => {
         email: 'local@example.com',
       },
     });
+  });
+});
+
+describe('registrationController - invite consumption', () => {
+  const invite = { token: 'hashed-invite', email: 'invitee@example.com' };
+
+  const buildRes = () => {
+    const res = {};
+    res.status = jest.fn(() => res);
+    res.send = jest.fn(() => res);
+    res.json = jest.fn(() => res);
+    return res;
+  };
+
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('consumes the invite once the account exists', async () => {
+    registerUser.mockResolvedValue({ status: 200, message: 'ok', userCreated: true });
+
+    await registrationController({ body: {}, invite }, buildRes());
+
+    expect(deleteTokens).toHaveBeenCalledWith({ token: 'hashed-invite' });
+  });
+
+  it('leaves the invite when registration is rejected', () => {
+    /** A mistyped password confirmation is the common case; it has to stay retryable. */
+    registerUser.mockResolvedValue({ status: 404, message: 'The passwords did not match' });
+
+    return registrationController({ body: {}, invite }, buildRes()).then(() => {
+      expect(deleteTokens).not.toHaveBeenCalled();
+    });
+  });
+
+  it('leaves the invite when the email is already in use, despite the 200', async () => {
+    /** `registerUser` returns the same status and message whether it created an account
+     *  or found the email taken, so the status alone cannot drive this decision. */
+    registerUser.mockResolvedValue({ status: 200, message: 'ok' });
+
+    await registrationController({ body: {}, invite }, buildRes());
+
+    expect(deleteTokens).not.toHaveBeenCalled();
+  });
+
+  it('does not attempt a deletion for an uninvited registration', async () => {
+    registerUser.mockResolvedValue({ status: 200, message: 'ok', userCreated: true });
+
+    await registrationController({ body: {} }, buildRes());
+
+    expect(deleteTokens).not.toHaveBeenCalled();
+  });
+
+  it('still reports success when consuming the invite fails', async () => {
+    /** The account exists by this point; reporting failure would be worse than
+     *  leaving a usable invite behind. */
+    registerUser.mockResolvedValue({ status: 200, message: 'ok', userCreated: true });
+    deleteTokens.mockRejectedValue(new Error('mongo unavailable'));
+    const res = buildRes();
+
+    await registrationController({ body: {}, invite }, res);
+
+    expect(res.status).toHaveBeenCalledWith(200);
+    expect(res.send).toHaveBeenCalledWith({ message: 'ok' });
+    expect(logger.error).toHaveBeenCalled();
+  });
+
+  it('never forwards the creation signal to the client', async () => {
+    registerUser.mockResolvedValue({ status: 200, message: 'ok', userCreated: true });
+    const res = buildRes();
+
+    await registrationController({ body: {}, invite }, res);
+
+    expect(res.send).toHaveBeenCalledWith({ message: 'ok' });
   });
 });
